remove numeric optimizations and move to BooleanSimplification
