+++ pom.xml
@@ -824,11 +824,6 @@
         <artifactId>jackson-mapper-asl</artifactId>
         <version>1.8.8</version>
       </dependency>
-      <dependency>
-        <groupId>org.spire-math</groupId>
-        <artifactId>spire_2.10</artifactId>
-        <version>0.9.0</version>
-      </dependency>
     </dependencies>
   </dependencyManagement>
