+++ sql/catalyst/pom.xml
@@ -44,10 +44,7 @@
       <groupId>org.scala-lang</groupId>
       <artifactId>scala-reflect</artifactId>
     </dependency>
-    <dependency>
-      <groupId>org.spire-math</groupId>
-      <artifactId>spire_2.10</artifactId>
-    </dependency>
+    
     <dependency>
       <groupId>org.apache.spark</groupId>
       <artifactId>spark-core_${scala.binary.version}</artifactId>
